Revert population vector init

+++ src/core/algorithms/nar/des/rng.h
@@ -6,7 +6,7 @@ namespace algos::des {
 
 class RNG {
 private:
-    int const kSeed = 2;
+    long unsigned const kSeed = 2;
     std::mt19937 rng_{kSeed};
     std::uniform_real_distribution<double> uni_{0.0, 1.0};
